+++ src/main/java/net/detalk/api/infrastructure/alarm/discord/DiscordAlarmSender.java
@@ -3,8 +3,8 @@ package net.detalk.api.infrastructure.alarm.discord;
 import java.util.EnumSet;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import net.detalk.api.infrastructure.alarm.AlarmErrorMessage;
-import net.detalk.api.infrastructure.alarm.AlarmSender;
+import net.detalk.api.alarm.domain.AlarmErrorMessage;
+import net.detalk.api.alarm.service.AlarmSender;
 import net.detalk.api.support.EnvironmentHolder;
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.JDABuilder;
