+++ src/main/java/net/detalk/api/service/DetalkInitializer.java
@@ -8,7 +8,7 @@ import java.util.Set;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import net.detalk.api.infrastructure.alarm.AlarmSender;
+import net.detalk.api.alarm.service.AlarmSender;
 import net.detalk.api.role.domain.Role;
 import net.detalk.api.plan.service.PricingPlanCache;
 import net.detalk.api.role.repository.RoleRepository;
