+++ src/main/java/net/detalk/api/support/JooqConfig.java
@@ -1,7 +1,7 @@
 package net.detalk.api.support;
 
 import lombok.RequiredArgsConstructor;
-import net.detalk.api.infrastructure.alarm.AlarmSender;
+import net.detalk.api.alarm.service.AlarmSender;
 import net.detalk.api.support.listener.PerformanceListener;
 import org.springframework.boot.autoconfigure.jooq.DefaultConfigurationCustomizer;
 import org.springframework.context.annotation.Bean;
