+++ src/main/java/net/detalk/api/support/WebExceptionHandler.java
@@ -3,8 +3,8 @@ package net.detalk.api.support;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.Arrays;
 import lombok.extern.slf4j.Slf4j;
-import net.detalk.api.infrastructure.alarm.AlarmErrorMessage;
-import net.detalk.api.infrastructure.alarm.AlarmSender;
+import net.detalk.api.alarm.domain.AlarmErrorMessage;
+import net.detalk.api.alarm.service.AlarmSender;
 import net.detalk.api.support.error.ApiException;
 import net.detalk.api.support.error.ErrorCode;
 import net.detalk.api.support.error.ErrorMessage;
